% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halpal.R
\name{halpal}
\alias{halpal}
\alias{scale_colour_halpal}
\alias{scale_color_halpal}
\alias{scale_fill_halpal}
\title{Color palettes using Halton sequences in CIE L*a*b* color space}
\usage{
halpal(n = 20, L_min = 45, L_max = 80, a_min = -60, a_max = 80,
  b_min = -60, b_max = 80)

scale_colour_halpal(...)

scale_color_halpal(...)

scale_fill_halpal(...)
}
\arguments{
\item{n}{The number of colors to include in the palette.}

\item{L_min}{The minimum value of L in the CIE L*a*b* color space.}

\item{L_max}{The maximum value of L in the CIE L*a*b* color space.}

\item{a_min}{The minimum value of a in the CIE L*a*b* color space.}

\item{a_max}{The maximum value of a in the CIE L*a*b* color space.}

\item{b_min}{The minimum value of b in the CIE L*a*b* color space.}

\item{b_max}{The maximum value of b in the CIE L*a*b* color space.}

\item{...}{Further arguments passed to `halpal`.}
}
\value{
A vector of colors.
}
\description{
Color palettes using Halton sequences in CIE L*a*b* color space
}
\examples{
n <- 20
barplot(rep(1, n), col=tail(halpal(n+15), n), space=0, border=NA, axes=FALSE)

n <- 20
set.seed(123)
intercept <- runif(n, -5, 5)
slope <- -(intercept/5)*runif(n, 0, 1)
dat <- expand.grid(ID=paste0("ID:", 1:n), x=1:10)
dat$y <- intercept + dat$x*slope + runif(nrow(dat), 0, 3)
ggplot(dat, aes(x=x, y=y, color=ID)) + geom_line(size=1.3) +
  scale_color_halpal() + theme_bw()

}
